created logo and icons for logo
