--- src/components/Header/NavBar/style.css
@@ -1,3 +1,0 @@
-.app-bar {
-    background-color: rgb(206,219,48);
-}
