+++ src/components/LandingPage/Advantages/index.js
