+++ src/components/LandingPage/Footer/index.js
@@ -1,0 +1,14 @@
+import React from "react";
+import "./style.css";
+
+function Footer() {
+  return (
+    <footer className="page-footer font-small blue pt-4">
+      <div className="card-footer text-muted">
+        <a href="/"> our footer</a>
+      </div>
+    </footer>
+  );
+}
+
+export default Footer;
