+++ src/components/LandingPage/Footer/style.css
@@ -1,0 +1,8 @@
+.page-footer {
+  background-color: "grey";
+  width: "100%";
+}
+
+a {
+  color: "black";
+}
