+++ src/components/LandingPage/Header/style.css
@@ -1,22 +1,23 @@
 .jumbotron-fluid {
   color: rgb(98, 250, 131);
   font-family: "Didact Gothic", sans-serif;
-  background-image: url(./backgroundHeader.jpeg);
+  background-image: url(../images/header.jpg);
   background-repeat: no-repeat;
   background-size: cover;
-  width: 100%;
+  width: 90%;
+  height: 500px;
   margin: auto;
   margin-top: 5px;
   margin-left: 5px;
   margin-right: 5px;
-  color: rgb(20, 12, 2);
-  border-radius: 15px;
+  color: rgb(253, 252, 250);
+  text-align: center;
 }
 
 h1 {
   font-family: "Josefin Sans", sans-serif;
   font-size: 80px;
-  color: rgb(41, 41, 26);
+  color: rgb(250, 250, 244);
   font-weight: bold;
 }
