+++ src/components/LandingPage/index.js
@@ -1,0 +1,12 @@
+import React from "react";
+import NavBar from "./NavBar";
+import Header from "./Header";
+
+export default function LandingPage() {
+  return (
+    <>
+      <NavBar />
+      <Header />
+    </>
+  );
+}
